chore(tests): delete duplicate fixtures file and update references
