+++ src/__tests__/shared/fixtures.ts
@@ -1,4 +1,9 @@
+import { capitalize } from 'lodash';
+import { WordDialect } from '../../types/word';
+import { NextFunction, Request, Response } from 'express';
+import { Flags } from '../../controllers/utils/types';
 import LanguageEnum from '../../shared/constants/LanguageEnum';
+import WordAttributeEnum from '../../shared/constants/WordAttributeEnum';
 import { SuggestionSourceEnum } from '../../shared/constants/SuggestionSourceEnum';
 import WordClass from '../../shared/constants/WordClass';
 import {
@@ -7,15 +12,21 @@ import {
   IncomingWord,
   DeveloperDocument,
   DeveloperUsage,
+  OutgoingExample,
+  OutgoingWord,
 } from '../../types';
 import { Types } from 'mongoose';
 import AccountStatus from '../../shared/constants/AccountStatus';
 import ApiType from '../../shared/constants/ApiType';
 import Plan from '../../shared/constants/Plan';
 
+interface RequestOptions {
+  noAuthorizationHeader?: boolean;
+}
+
 export const documentId = new Types.ObjectId('569ed8269353e9f4c51617aa');
 
-export const wordFixture = (wordData: Partial<IncomingWord>) => ({
+export const incomingWordFixture = (wordData: Partial<IncomingWord>) => ({
   definitions: [],
   dialects: [],
   tags: [],
@@ -54,7 +65,7 @@ export const definitionFixture = (definitionData: Partial<Definition>) => ({
   ...definitionData,
 });
 
-export const exampleFixture = (exampleData: Partial<IncomingExample>) => ({
+export const incomingExampleFixture = (exampleData: Partial<IncomingExample>) => ({
   source: { text: '', language: LanguageEnum.UNSPECIFIED, pronunciations: [] },
   translations: [{ text: '', language: LanguageEnum.UNSPECIFIED, pronunciations: [] }],
   meaning: '',
@@ -94,3 +105,86 @@ export const developerUsageFixture = (developerFixture: Partial<DeveloperUsage>)
   },
   ...developerFixture,
 });
+
+export const outgoingWordFixture = (data?: Partial<OutgoingWord>) => ({
+  attributes: Object.values(WordAttributeEnum).reduce(
+    (finalAttributes, attribute) => ({ ...finalAttributes, [attribute]: false }),
+    {}
+  ),
+  conceptualWord: '',
+  frequency: 1,
+  hypernyms: [],
+  hyponyms: [],
+  pronunciation: '',
+  relatedTerms: [],
+  stems: [],
+  updatedAt: new Date(),
+  variations: [],
+  word: '',
+  wordPronunciation: '',
+  definitions: [],
+  dialects: [],
+  tags: [],
+  id: `${new Types.ObjectId()}`,
+  ...data,
+});
+
+export const outgoingExampleFixture = (data?: Partial<OutgoingExample>) => ({
+  id: `${new Types.ObjectId()}`,
+  associatedDefinitionsSchemas: [],
+  associatedWords: [],
+  source: { text: '', language: LanguageEnum.UNSPECIFIED, pronunciations: [] },
+  translations: [{ text: '', language: LanguageEnum.UNSPECIFIED, pronunciations: [] }],
+  meaning: '',
+  nsibidi: '',
+  nsibidiCharacters: [],
+  pronunciations: [],
+  origin: SuggestionSourceEnum.INTERNAL,
+  updatedAt: new Date(),
+  ...data,
+});
+
+export const dialectFixture = (data?: Partial<WordDialect>) => ({
+  dialects: [],
+  id: `${new Types.ObjectId()}`,
+  pronunciation: '',
+  variations: [],
+  word: '',
+  ...data,
+});
+
+export const flagsFixture = (data?: Partial<Flags>) => ({
+  examples: false,
+  dialects: false,
+  resolve: false,
+  style: '',
+  ...data,
+});
+
+export const requestFixture = (
+  {
+    body = {},
+    params = {},
+    headers = {},
+  }: {
+    body?: { [key: string]: string },
+    params?: { [key: string]: string },
+    headers?: { [key: string]: string },
+  } = {},
+  options?: RequestOptions
+): Request => ({
+  body,
+  params,
+  headers,
+  query: {},
+  // @ts-expect-error get
+  get: (header: string) => headers[header] || headers[capitalize(header)],
+});
+export const statusSendMock = jest.fn();
+export const responseFixture = (): Response => ({
+  // @ts-expect-error status
+  status: jest.fn(() => ({ send: statusSendMock })),
+  send: jest.fn(),
+  redirect: jest.fn(),
+});
+export const nextFunctionFixture = (): NextFunction => jest.fn();
