+++ src/controllers/__tests__/developers.test.ts
@@ -2,7 +2,7 @@ import {
   requestFixture,
   responseFixture,
   nextFunctionFixture,
-} from '../../../__tests__/shared/fixtures';
+} from '../../__tests__/shared/fixtures';
 import {
   getDeveloper,
   getDeveloperByEmail,
