+++ src/controllers/__tests__/examples.test.ts
@@ -1,11 +1,11 @@
-import { exampleFixture } from '../../__tests__/shared/fixtures';
+import { incomingExampleFixture } from '../../__tests__/shared/fixtures';
 import LanguageEnum from '../../shared/constants/LanguageEnum';
 import { SuggestionSourceEnum } from '../../shared/constants/SuggestionSourceEnum';
 import { convertToV1Example } from '../examples';
 
 describe('examples', () => {
   it('converts example pronunciations to pronunciation for v1', () => {
-    const example = exampleFixture({
+    const example = incomingExampleFixture({
       source: {
         text: 'igbo',
         language: LanguageEnum.IGBO,
