+++ src/controllers/utils/__tests__/minimizeVerbsAndSuffixes.test.ts
@@ -1,27 +1,27 @@
 import Version from '../../../shared/constants/Version';
 import WordClass from '../../../shared/constants/WordClass';
-import { definitionFixture, wordFixture } from '../../../__tests__/shared/fixtures';
+import { definitionFixture, incomingWordFixture } from '../../../__tests__/shared/fixtures';
 import minimizeVerbsAndSuffixes from '../minimizeVerbsAndSuffixes';
 import WordClassEnum from '../../../shared/constants/WordClassEnum';
 
 describe('minimizeVerbsAndSuffixes', () => {
   it('minimizes the verbs and suffixes to include basic fields', () => {
     const words = [
-      wordFixture({
+      incomingWordFixture({
         word: 'first word',
         definitions: [definitionFixture({})],
         stems: [],
         relatedTerms: [],
         id: '123',
       }),
-      wordFixture({
+      incomingWordFixture({
         word: 'second word',
         definitions: [definitionFixture({ wordClass: WordClassEnum.ADV })],
         stems: [],
         relatedTerms: [],
         id: '456',
       }),
-      wordFixture({
+      incomingWordFixture({
         word: 'third word',
         definitions: [definitionFixture({ wordClass: WordClassEnum.PREP })],
         stems: [],
